src/main/java/com/zm/context/annotation/ClassPathBeanDefinitionScanner.java
package com.zm.context.annotation;

import com.zm.beans.BeanDefinition;
import com.zm.beans.factory.BeanDefinitionStoreException;
import com.zm.beans.factory.support.BeanDefinitionRegistry;
import com.zm.beans.factory.support.BeanNameGenerator;
import com.zm.core.io.Resource;
import com.zm.core.io.support.PackageResourceLoader;
import com.zm.core.type.classreading.MetadataReader;
import com.zm.core.type.classreading.SimpleMetadataReader;
import com.zm.stereotype.Component;
import org.springframework.util.StringUtils;

import java.io.IOException;
import java.util.LinkedHashSet;
import java.util.Set;

//扫描注解的方式获取BeanDefinition
public class ClassPathBeanDefinitionScanner {


    private final BeanDefinitionRegistry registry;

    //一个包变成resource[]
    private PackageResourceLoader resourceLoader = new PackageResourceLoader();

    private BeanNameGenerator beanNameGenerator = new AnnotationBeanNameGenerator();

    public ClassPathBeanDefinitionScanner(BeanDefinitionRegistry registry) {
        this.registry = registry;
    }

    public Set<BeanDefinition> doScan(String packagesToScan) {

        //多个目录
        String[] basePackages = StringUtils.tokenizeToStringArray(packagesToScan, ",");

        Set<BeanDefinition> beanDefinitions = new LinkedHashSet<>();
        for (String basePackage : basePackages) {
            Set<BeanDefinition> candidates = findCandidateComponents(basePackage);
            for (BeanDefinition candidate : candidates) {
                beanDefinitions.add(candidate);
                registry.registryBeanDefinition(candidate.getID(), candidate);

            }
        }
        return beanDefinitions;
    }

    //加载一个包下的注解bean
    public Set<BeanDefinition> findCandidateComponents(String basePackage) {
        Set<BeanDefinition> candidates = new LinkedHashSet<>();
        try {

            Resource[] resources = this.resourceLoader.getResources(basePackage);
            for (Resource resource : resources) {
                try {
                    MetadataReader metadataReader = new SimpleMetadataReader(resource);
                    if (metadataReader.getAnnotationMetadata().hasAnnotation(Component.class.getName())) {
                        ScannedGenericBeanDefinition sbd = new ScannedGenericBeanDefinition(metadataReader.getAnnotationMetadata());
                        String beanName = this.beanNameGenerator.generateBeanName(sbd, this.registry);
                        sbd.setId(beanName);
                        candidates.add(sbd);
                    }
                } catch (Throwable ex) {
                    throw new BeanDefinitionStoreException(
                            "Failed to read candidate component class: " + resource, ex);
                }

            }
        } catch (IOException ex) {
            throw new BeanDefinitionStoreException("I/O failure during classpath scanning", ex);
        }
        return candidates;
    }

}